Make sure method `handles` trait is not re-applied (#4)
rakudo/rakudo#4994 resulted in an unpleased issue where duplicating the
same method object under a different name caused lazy `handles`
application to be repeated and delegation methods re-added again. For
now I have added a check which prevents this to happen if the method
duplication is detected, but it's an additional loop iterating over
already installed ones.

To fix this issue I have added `:$handles` parameter to `add_method`.
Currently it would serve as a mere optimization if negated. But in the
future (likely after this commit makes it into `Method::Also` release)
I plan to remove the loop and make any third party module to make sure
they don't cause delegation duplication.